log_back/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
mysql-connector-python==8.3.0
pyjwt==2.8.0
passlib==1.7.4
python-dotenv==1.0.1
email-validator==2.3.0